Support extraction of $gettext from vue <script> tags
The goal of this commit is to introduce the possibility
to extract localized strings located in Vue.js <script/>
tags. This is a first step. For the moment, only strings
marked as translatable using the $gettext function are
extracted.

How to test:
- Setup a .vue file like:
    <template>
        <h1>{{ greeting_message }}</h1>
    </template>
    <script>
        export default {
            name: "greetings",
            computed: {
                greeting_message() {
                    return this.$gettext("Hello there!")
                },
                duplicated_greeting_message() {
                    return  this.$gettext("Hello there!")
                }
            }
        }
     </script>

- Run compile-cli.js with this file

Expected result -> Strings present in the script part of the vue
component have been successfully extracted.