extract: Fix parsing HTML
Commit dc49c36adeb458ea48e1d5a4f965e1c75a83d674 introduced a subtle bug:
while the previous implementation of extract-cli would always override
`data` with the returned value of `preprocessScriptTags`, leading to an
empty value in case of an error, the new implementation didn't do so and
left it as-is. Then, it passed the data to `parseJavascript`,
effectively trying to parse HTML using acorn, which crashes.

This commit moves logic out from extract-cli.js back into extract.js,
using a similar function composition for Javascript as there exists for
templates. The behaviour should be the same, except now we effectively
return an empty string to be passed to parseJavascript, and not the raw
HTML.